Add some tests for `compare`.
This patch adds two passing tests and one failing test for the `compare`
function. The failing test fails because the total length of the samples
(summed for the two benchmarks being compared) exceeds 30.

This hits the following code in `compare`:

  // reject the null hyphothesis the two samples come from the
  // same population (i.e. have the same median) if...
  if (size1 + size2 > 30) {
    // ...the z-stat is greater than 1.96 or less than -1.96
    // http://www.statisticslectures.com/topics/mannwhitneyu/
    zStat = getZ(u);
    return abs(zStat) > 1.96 ? (zStat > 0 ? -1 : 1) : 0;
  }

This seems to always return `1`, i.e. that the `other` benchmark is
faster.