Add human-readable manifest.txt to nano.jar
It is JAR file manifest attributes for security, which is about the permissions
of the Java applet. If it is unset, user is prompted to run the applet or not.

We have successfully set the manifest in binary form, as in
META-INF/MANIFEST.MF in nano.jar. However, its source code of the manifest was
not committed. I add the source to ensure that we can regenerate nano.jar and
to conform free software license.