[Patch v2] Fix Benchmark#toString() stringification of error
Benchmark#toString() calls join() to convert its error property to a
string.  Since the specified properties of Error (message and name) are
not enumerable, when error is an instance of Error this produces a
string which is either empty or contains only properties assigned to the
error after instantiation.  It appears that this has been the case since
its introduction in ee59f95.

This commit attempts to achieve the intended behavior by calling join on
an object with the name and message properties of the error in addition
to any enumerable properties.

Changes in v2:
* Replace ternary operators with if/else blocks.
* Replace typeof comparison to 'object' with _.isObject.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>