Fix benchmark tests when running with minified library code
This fixes an issue I found when using benchmark and my library was minified. I don't really understand all of the implications, but what seems to have happened is that the line "deferred.suResolve()" errors as "deferred" was undefined in minified.

By inspection, I noted that similar code was using the "interpolate" function and replaced "#d" with the name of the deferred variable.